requirements.txt
annotated-types==0.7.0
anyio==4.11.0
CacheControl==0.14.3
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cryptography==46.0.3
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.119.0
firebase_admin==7.1.0
google-api-core==2.26.0
google-api-python-client==2.185.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.4.1
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.75.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==1.0.0
langgraph==1.0.0
langgraph-checkpoint==2.1.2
langgraph-prebuilt==1.0.0
langgraph-sdk==0.2.9
langsmith==0.4.37
msgpack==1.1.2
oauthlib==3.3.1
orjson==3.11.3
ormsgpack==1.11.0
packaging==25.0
proto-plus==1.26.1
protobuf==6.33.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.2
pydantic_core==2.41.4
pygraphviz==1.14
PyJWT==2.10.1
pyparsing==3.2.5
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9.1
sniffio==1.3.1
starlette==0.48.0
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.37.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
xxhash==3.6.0
zstandard==0.25.0
